[mametool] fixed bug with mame0.211 softlists...
...mametool reported that it tried to use a softlist for which there was
no hasfile on disk. this was in the softlist other-game-name
functionality, where instead of passing the results of filtering
softlists for has existence and other matters, there was another test
against softlist existence. Core issue was not passing the results of
first filtering downstream, probably because R.map seems unable to pass
on its list without a helper function, which made it difficult to see how
to achieve that

there's a cleanup operation to do, since softlistParams was being used
for a lot of similar info, shouldn't really be passing both 'list' and
  many of the details of softlistParams to various fns