[mametool] proper fix for the briefname issue address 2 commits ago...
...the root cause of briefname not appearing was that in the MAME DTD, a
device doesn't have to have an 'instance', which means it gets no name
or briefname. A device without an instance is useless to us, and noticed
that we were still printing efind entries with 'undefined' devices
without this fix. Devices without extensions were all the same story,
also returning null previously for these was an unnecessary mistake,
a device that doesn't have a name or filetypes it accepts is useless to
us here