[pkg] implement least-preferred plan for exe build...
see https://github.com/vercel/pkg/issues/1319 <-- tried making a pkg
property in package.json specifying assets, resolving prebuild-install
for npm to no avail. There's no way to stop pkg packaging the macos
binary, so patch in the windows iconv node file before build, and
put the macos iconv node file back again once compiling completes (for
dev). There is another node file for expat, but this doesn't pick up and
the old behaviour of expecting it in the destination folder is
maintained - is this due to an obfuscated require? If so, that could be
a more elegant sloution?