[mame] remove string similarity print paths code...
because rompaths are arbitrary, we now use a UI to prompt the user to
fill in rompath types, we don't try and automatically work out path
types from the mame ini's mame rompath setting. So remove all the code
that tried to do that, it was wrong...

also remove tests, but add tests for getting rompaths. Can't really test
determinePathToMameIni because its too impure