[mame] all inputs to inputs folder...
some inputs are in root not inputs for historical reasons, the inputs
folder was not in source control. So the qp settings file and mame.ini
now in inputs (or rather inputs/current).

There seems no use for the mame.ini atm however, it should be part of
a getRomPath integration test...